optimised

+++ server/controllers/productController.js
@@ -1,8 +1,21 @@
-const Product = require('../models/productModel')
+const Product = require('../models/productModel');
+
+const getProducts = async (req, res) => {
+    try {
+        const products = await Product.find();
+        return res.status(200).send(products);
+    } catch (err) {
+        return res.status(500).json({error: err.message});
+    }
+};
 
 const addProduct = async (req, res) => {
     try {
         const { name, price, description } = req.body;
+
+        if (!name || !price) {
+            return res.status(400).json({error: "Please provide all fields!"});
+        };
 
         const newProduct = new Product({
             name,
@@ -11,6 +24,7 @@ const addProduct = async (req, res) => {
         });
 
         await newProduct.save();
+
         return res.status(201).json({ message: "Product Added Successfully!" });
     } catch (err) {
         return res.status(500).json({ error: err.message });
@@ -21,6 +35,10 @@ const updateProduct = async (req, res) => {
     try {
         const { id, name, price, description } = req.body;
 
+        if (!id || !name || !price) {
+            return res.status(400).json({error: "Please check the constraints!"});
+        };
+
         const updatedProduct = await Product.findByIdAndUpdate(id, {
             name,
             price,
@@ -29,7 +47,7 @@ const updateProduct = async (req, res) => {
 
         if (!updatedProduct) {
             return res.status(404).json({ message: "Product not found" });
-        }
+        };
 
         return res.status(200).json({ message: "Product Updated Successfully!", product: updatedProduct });
     } catch (err) {
@@ -40,6 +58,10 @@ const updateProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
     try {
         const { id } = req.params;
+
+        if (!id) {
+            return res.status(400).json({error: "Please provide ID!"});
+        };
 
         const deletedProduct = await Product.findByIdAndDelete(id);
 
@@ -53,7 +75,4 @@ const deleteProduct = async (req, res) => {
     }
 };
 
-
-
-
-module.exports = { addProduct, updateProduct, deleteProduct };
+module.exports = { getProducts, addProduct, updateProduct, deleteProduct };
